Set Skeleton component to maximum width
The changes provide a more uniform and sleeker design by adjusting the Skeleton component in the UI package to have a maximum width of '100%'. This modification ensures the component's flexibility and adaptability across a wide variety of screen sizes, promoting better responsiveness of the user interface. Version of UI package is also updated to 0.1.68.